# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/jobs/{id}/purchase-orders/revisions": {
      "post": {
        "operationId": "PurchaseOrderController_revise",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": "JOB_NOT_LIVE: only live jobs accept watchdog inputs. IDEMPOTENCY_CONFLICT: command identity was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "JOB_NOT_LIVE",
                        "IDEMPOTENCY_CONFLICT"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Append a synthetic purchase-order draft revision",
        "tags": [
          "purchase-orders"
        ]
      }
    },
    "/jobs/{id}/purchase-orders/placement": {
      "post": {
        "operationId": "PurchaseOrderController_place",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": "JOB_NOT_LIVE: only live jobs accept watchdog inputs. IDEMPOTENCY_CONFLICT: command identity was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "JOB_NOT_LIVE",
                        "IDEMPOTENCY_CONFLICT"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Authorize one exact simulated purchase-order placement",
        "tags": [
          "purchase-orders"
        ]
      }
    },
...
    "/jobs/{id}/supplier-matches": {
      "get": {
        "operationId": "SupplierMatchController_view",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read exact order, receipt and bill match revisions",
        "tags": [
          "supplier-matches"
        ]
      },
      "post": {
        "operationId": "SupplierMatchController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": "JOB_NOT_LIVE: only live jobs accept watchdog inputs. IDEMPOTENCY_CONFLICT: command identity was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "JOB_NOT_LIVE",
                        "IDEMPOTENCY_CONFLICT"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Create a deterministic source match proposal",
        "tags": [
          "supplier-matches"
        ]
      }
    },
...
    "/jobs/{id}/things-to-check/evaluate": {
      "post": {
        "operationId": "ThingsToCheckController_evaluate",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": "JOB_NOT_LIVE: only live jobs accept watchdog inputs. IDEMPOTENCY_CONFLICT: command identity was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "JOB_NOT_LIVE",
                        "IDEMPOTENCY_CONFLICT"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Evaluate confirmed matched facts with a versioned pure rule",
        "tags": [
          "things-to-check"
        ]
      }
    },
    "/jobs/{id}/things-to-check/review": {
      "post": {
        "operationId": "ThingsToCheckController_review",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": "JOB_NOT_LIVE: only live jobs accept watchdog inputs. IDEMPOTENCY_CONFLICT: command identity was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "JOB_NOT_LIVE",
                        "IDEMPOTENCY_CONFLICT"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Dismiss or dispute a finding with a reason",
        "tags": [
          "things-to-check"
        ]
      }
    },
    "/jobs/{id}/things-to-check/supersede-bill": {
      "post": {
        "operationId": "ThingsToCheckController_supersede",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": "JOB_NOT_LIVE: only live jobs accept watchdog inputs. IDEMPOTENCY_CONFLICT: command identity was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "JOB_NOT_LIVE",
                        "IDEMPOTENCY_CONFLICT"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Approve a confirmed source-backed replacement bill",
        "tags": [
          "things-to-check"
        ]
      }
    },
...
    "/jobs/{id}/relevance-inbox/seed": {
      "post": {
        "operationId": "InboxRelevanceController_seed",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": "JOB_NOT_LIVE: only live jobs accept watchdog inputs. IDEMPOTENCY_CONFLICT: command identity was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "JOB_NOT_LIVE",
                        "IDEMPOTENCY_CONFLICT"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Record deterministic synthetic findings",
        "tags": [
          "inbox relevance"
        ]
      }
    },
...
    "/jobs/{id}/relevance-inbox/{decisionId}/dismiss": {
      "post": {
        "operationId": "InboxRelevanceController_dismiss",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": "JOB_NOT_LIVE: only live jobs accept watchdog inputs. IDEMPOTENCY_CONFLICT: command identity was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "JOB_NOT_LIVE",
                        "IDEMPOTENCY_CONFLICT"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Dismiss one job-bound advisory Decision idempotently",
        "tags": [
          "inbox relevance"
        ]
      }
    },
...
  },
  "info": {
    "title": "JobGuard API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}