# apps/api/src/inbox-relevance.controller.ts
import { practiceCookie } from "./practice-access.js";
import{Headers,Body,Controller,Get,Param,Post}from"@nestjs/common";import{ApiOperation,ApiTags}from"@nestjs/swagger";import{Pool}from"pg";import{InboxRelevanceApplication}from"./inbox-relevance.application.js";@ApiTags("inbox relevance")@Controller("jobs/:id/relevance-inbox")export class InboxRelevanceController{constructor(private readonly pool:Pool){}@Get()@ApiOperation({summary:"List source-linked mandatory and advisory Decisions"})view(@Param("id")id:string, @Headers("cookie") cookie?:string){return new InboxRelevanceApplication(this.pool,practiceCookie(cookie)).view(id)}@Post("seed")@ApiOperation({summary:"Record deterministic synthetic findings"})seed(@Param("id")id:string,@Body()body:unknown, @Headers("cookie") cookie?:string){return new InboxRelevanceApplication(this.pool,practiceCookie(cookie)).seed(id,body)}@Post("budget")@ApiOperation({summary:"Record an advisory budget revision"})budget(@Param("id")id:string,@Body()body:unknown, @Headers("cookie") cookie?:string){return new InboxRelevanceApplication(this.pool,practiceCookie(cookie)).budget(id,body)}@Post(":decisionId/dismiss")@ApiOperation({summary:"Dismiss one job-bound advisory Decision idempotently"})dismiss(@Param("id")id:string,@Param("decisionId")decisionId:string,@Body()body:unknown, @Headers("cookie") cookie?:string){return new InboxRelevanceApplication(this.pool,practiceCookie(cookie)).dismiss(id,decisionId,body)}}