# apps/api/src/purchase-order.controller.ts
import { practiceCookie } from "./practice-access.js";
import{Headers,Body,Controller,Get,Param,Post}from"@nestjs/common";import{ApiOperation,ApiTags}from"@nestjs/swagger";import{Pool}from"pg";import{PurchaseOrderApplication}from"./purchase-order.application.js";
@ApiTags("purchase-orders")@Controller("jobs/:id/purchase-orders")export class PurchaseOrderController{constructor(private readonly pool:Pool){}@Get()@ApiOperation({summary:"Read the authoritative synthetic purchase-order pre-commit state"})get(@Param("id")id:string, @Headers("cookie") cookie?:string){return new PurchaseOrderApplication(this.pool,practiceCookie(cookie)).view(id)}@Post("revisions")@ApiOperation({summary:"Append a synthetic purchase-order draft revision"})revise(@Param("id")id:string,@Body()body:unknown, @Headers("cookie") cookie?:string){return new PurchaseOrderApplication(this.pool,practiceCookie(cookie)).revise(id,body)}@Post("placement")@ApiOperation({summary:"Authorize one exact simulated purchase-order placement"})place(@Param("id")id:string,@Body()body:unknown, @Headers("cookie") cookie?:string){return new PurchaseOrderApplication(this.pool,practiceCookie(cookie)).place(id,body)}}