# apps/api/src/things-to-check.controller.ts
import { practiceCookie } from "./practice-access.js";
import{Headers,Body,Controller,Get,Param,Post}from"@nestjs/common";import{ApiOperation,ApiTags}from"@nestjs/swagger";import{Pool}from"pg";import{ThingsToCheckApplication}from"./things-to-check.application.js";@ApiTags("things-to-check")@Controller("jobs/:id/things-to-check")export class ThingsToCheckController{constructor(private readonly pool:Pool){}@Get()@ApiOperation({summary:"Read cited supplier discrepancy decisions"})view(@Param("id")id:string, @Headers("cookie") cookie?:string){return new ThingsToCheckApplication(this.pool,practiceCookie(cookie)).view(id)}@Post("evaluate")@ApiOperation({summary:"Evaluate confirmed matched facts with a versioned pure rule"})evaluate(@Param("id")id:string,@Body()body:unknown, @Headers("cookie") cookie?:string){return new ThingsToCheckApplication(this.pool,practiceCookie(cookie)).evaluate(id,body)}@Post("review")@ApiOperation({summary:"Dismiss or dispute a finding with a reason"})review(@Param("id")id:string,@Body()body:unknown, @Headers("cookie") cookie?:string){return new ThingsToCheckApplication(this.pool,practiceCookie(cookie)).review(id,body)}@Post("supersede-bill")@ApiOperation({summary:"Approve a confirmed source-backed replacement bill"})supersede(@Param("id")id:string,@Body()body:unknown, @Headers("cookie") cookie?:string){return new ThingsToCheckApplication(this.pool,practiceCookie(cookie)).supersede(id,body)}}