# apps/web/app/api/jobs/[id]/supplier-matches/route.ts
import { practiceFailure } from "../../../../lib/synthetic-server";
import{cookies}from"next/headers";import{NextResponse}from"next/server";import{hasSyntheticSession}from"../../../../lib/synthetic-server";import{workspaceApplication}from"../../../../lib/workspace-server";const auth=async()=>hasSyntheticSession((await cookies()).get("jg_session")?.value);export async function GET(_r:Request,{params}:{params:Promise<{id:string}>}){if(!await auth())return NextResponse.json({code:"UNAUTHENTICATED"},{status:401});try{return NextResponse.json(await (await workspaceApplication()).supplierMatches.view((await params).id))}catch(e){const denied=practiceFailure(e);if(denied)return denied;return NextResponse.json({code:e instanceof Error?e.message:"INVALID_MATCH"},{status:400})}}export async function POST(r:Request,{params}:{params:Promise<{id:string}>}){if(!await auth())return NextResponse.json({code:"UNAUTHENTICATED"},{status:401});try{return NextResponse.json(await (await workspaceApplication()).supplierMatches.create((await params).id,await r.json()))}catch(e){const denied=practiceFailure(e);if(denied)return denied;const code=e instanceof Error?e.message:"INVALID_MATCH";return NextResponse.json({code},{status:code.includes("CONFLICT")?409:400})}}